Drf_Project/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-debug-toolbar==4.4.6
django-filter==24.3
mysqlclient==2.2.6
social-auth-app-django==5.4.2
requests==2.32.3
requests-oauthlib==2.0.0
oauthlib==3.2.2
PyJWT==2.10.1
PyMySQL==1.1.1
sqlparse==0.5.2